Option Chain ABstraction defined and implemented with Stock/Indices Abstractions.
