+++ tests/test_nse/test_nse_config.py
@@ -111,7 +111,7 @@ def test_get_all_etfs(nse_config):
     "symbol", ["NIFTY", "NIFTYNXT50", "BANKNIFTY", "RELIANCE", "SBIN"]
 )
 def test_get_option_chain(nse_config, symbol):
-    response = nse_config.get_option_chain(symbol)
+    response = nse_config.get_option_chain_data(symbol)
     assert isinstance(response, dict)
     assert all(i in response.keys() for i in ("records", "filtered"))
