+++ trade/nse/data_generics.py
@@ -1,4 +1,6 @@
 from typing import Literal, Optional, TypeVar
+from trade.option_chain.index_option_chain import IndexOptionChainAnalysis
+from trade.option_chain.stock_option_chain import StockOptionChain
 
 
 class NSEDataGeneric:
@@ -10,11 +12,11 @@ class NSEDataGeneric:
         match oc_type:
 
             case "index":
-                return self.option_chain_result("IndexChainType")
+                return self.option_chain_result(IndexOptionChainAnalysis)
 
             case "equity":
                 return (
-                    self.option_chain_result("EquityOptionCHainType")
+                    self.option_chain_result(StockOptionChain)
                     if self.is_fno
                     else None
                 )
@@ -24,10 +26,8 @@ class NSEDataGeneric:
 
     def option_chain_result(self, OptionChain: type):
 
-        option_chain_data = self._nse_config.get_option_chain.get_option_chain(
-            self.symbol
-        )
+        option_chain_data = self._nse_config.get_option_chain_data.get_option_chain_data(self.symbol)
 
-        oc_obj = OptionChain(option_chain_data, type="")
+        oc_obj = OptionChain(symbol, option_chain_data, self.dated, strike_multiples)
 
-        return oc_obj
+        return oc_obj.option_chain_output()
